Merge to main

+++ src/main/java/com/palmatoro/cmmimplant/service/ProjectService.java
@@ -5,8 +5,8 @@ import java.util.Date;
 import java.util.List;
 
 import com.palmatoro.cmmimplant.domain.Project;
-import com.palmatoro.cmmimplant.domain.User;
 import com.palmatoro.cmmimplant.domain.Project.ProjectType;
+import com.palmatoro.cmmimplant.domain.User;
 import com.palmatoro.cmmimplant.exception.ResourceNotFoundException;
 import com.palmatoro.cmmimplant.repository.ProjectRepository;
 
@@ -66,7 +66,7 @@ public class ProjectService {
 
     @Transactional
     public Project editProject(Integer id, String name, ProjectType projectType, Date startDate, Date endDate){
-        Project project = projectRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("No project found on ID: " + id));
+        Project project = projectRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("No se ha encontrado un proyecto con el ID: " + id));
 
         project.setName(name);
         project.setProjectType(projectType);
@@ -77,8 +77,11 @@ public class ProjectService {
     }
 
     @Transactional
-    public void deleteProjectById(Integer id){
-        Project project = projectRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("No project found on ID: " + id));
+    public void deleteProjectById(Integer id) throws Exception {
+        Project project = projectRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("No se ha encontrado un proyecto con el ID: " + id));
+        if(getProjectByPrincipal().getId().equals(id)) {
+            throw new Exception("No puedes eliminar el proyecto al que perteneces.");
+        }
         projectRepository.delete(project);
     }
 
@@ -88,7 +91,7 @@ public class ProjectService {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String currentPrincipalName = authentication.getName();
 
-        User u = userService.getUserByEmail(currentPrincipalName);
+        User u = userService.getUserByUsername(currentPrincipalName);
 
         return u.getProject();
     }
